chore: adapt test code to be more readable and change configuration type
