+++ core/runtime/src/main/java/io/quarkiverse/jberet/runtime/QuarkusJobExecutor.java
@@ -14,18 +14,12 @@ class QuarkusJobExecutor extends JobExecutor {
     public QuarkusJobExecutor(Executor delegate, final ThreadPoolConfig threadPoolConfig, final JBeretConfig config) {
         super(delegate);
 
-        this.maxPoolSize = config.maxAsync().map(max -> {
-            try {
-                var maxAsync = Integer.parseInt(max);
-                if (maxAsync < 1) {
-                    Log.error("max-async value must be 1 or greater if set.");
-                    return null;
-                }
-                return maxAsync + 1; // Adapt for the fact that JBeret requires one thread for coordination.
-            } catch (NumberFormatException ex) {
-                Log.error("max-async configuration value must be numeric! Could not parse value: " + max + " to an integer.");
+        this.maxPoolSize = config.maxAsync().map(maxAsync -> {
+            if (maxAsync < 1) {
+                Log.error("max-async value must be 1 or greater if set.");
                 return null;
             }
+            return maxAsync + 1; // Adapt for the fact that JBeret requires one thread for coordination.
         }).orElse(threadPoolConfig.maxThreads.orElse(Math.max(8 * ProcessorInfo.availableProcessors(), 200)));
     }
